requirements.txt
# add every additional library we need to run the app here
streamlit
pandas
matplotlib.pyplot